+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - `at-rule-no-vendor-prefix`, `media-feature-name-no-vendor-prefix`, `property-no-vendor-prefix`, `selector-no-vendor-prefix`, `value-no-vendor-prefix` to encourage Autoprefixer usage
 - `string-quotes`, `selector-attribute-quotes`, `font-family-name-quotes` to standardise string usage
+- `unspecified: 'bottomAlphabetical'` to `order/properties-order` rule
 
 ## [0.1.2]
 ### Changed